+++ src/models/mod.rs
@@ -2,7 +2,6 @@ pub mod alignment;
 pub mod contig;
 pub mod contig_collection;
 pub mod cytoband;
-pub mod data;
 pub mod message;
 pub mod mode;
 pub mod reference;
