+++ src/states.rs
@@ -1,8 +1,11 @@
 use crate::error::TGVError;
+use crate::repository::{AlignmentRepository, AlignmentRepositoryEnum};
+
 use crate::models::{
     contig::Contig,
+    contig_collection::ContigCollection,
+
     cytoband::Cytoband,
-    data::Data,
     track::{feature::Gene, track::Track},
     message::{DataMessage, StateMessage},
     mode::InputMode,
@@ -13,10 +16,12 @@ use crate::models::{
         track_service::{
             TrackCache, TrackService, TrackServiceEnum, UcscApiTrackService, UcscDbTrackService,
         },
+
+        sequences::SequenceService
     },
     window::ViewingWindow,
 };
-use crate::settings::Settings;
+use crate::settings::{Settings, BackendType};
 use crate::traits::GenomeInterval;
 use crossterm::event::{KeyCode, KeyEvent};
 use ratatui::layout::Rect;
@@ -43,6 +48,23 @@ pub struct State {
 
     /// Error messages for display.
     pub errors: Vec<String>,
+
+
+    /// Alignment segments.
+    pub alignment: Option<Alignment>,
+
+    /// Tracks.
+    pub track: Option<Track<Gene>>,
+    pub track_cache: TrackCache,
+    
+
+    /// Sequences.
+    pub sequence: Option<Sequence>,
+    
+
+    // TODO: in the first implementation, refresh all data when the viewing window is near the boundary.
+    /// Contigs in the BAM header
+    pub contigs: ContigCollection,
 }
 
 /// Basics
@@ -62,6 +84,14 @@ impl State {
 
             settings,
             errors: Vec::new(),
+
+            alignment: None,
+            track: None,
+            track_cache: TrackCache::new(),
+            sequence: None,
+            contigs: ContigCollection::new(settings.reference.clone()) // TODO: load this
+
+            
         })
     }
 
@@ -187,16 +217,62 @@ impl State {
         }
     }
 
+
     pub fn track_checked(&self) -> Result<&Track<Gene>, TGVError> {
         self.data.track_checked()
     }
 }
 
 
-struct StateHandler {}
+pub struct StateHandler {
+    alignment_repository: AlignmentRepositoryEnum,
+
+    track_service: Option<TrackServiceEnum>,
+
+    sequence_service: Option<SequenceService>,
+
+
+
+}
 
 
 impl StateHandler {
+
+    pub async fn new(settings: &Settings) -> Result<Self, TGVError> {
+
+        let alignment_repository = AlignmentRepositoryEnum::from(settings)?;
+
+        let (track_service, sequence_service): (Option<TrackServiceEnum>, Option<SequenceService>) =
+            match settings.reference.as_ref() {
+                Some(reference) => {
+                    let ts = match settings.backend {
+                        BackendType::Api => TrackServiceEnum::Api(UcscApiTrackService::new()?),
+                        BackendType::Db => {
+                            TrackServiceEnum::Db(UcscDbTrackService::new(reference).await?)
+                        }
+                    };
+                    let ss = SequenceService::new(reference.clone())?;
+                    (Some(ts), Some(ss))
+                }
+                None => (None, None),
+            };
+        
+        Ok(Self {
+            alignment_repository,
+            track_service,
+            sequence_service
+        })
+    }
+
+    pub async fn close(&mut self) -> Result<(), TGVError> {
+        if let Some(ts) = self.track_service.as_mut() {
+            ts.close().await?;
+        }
+        if let Some(ss) = self.sequence_service.as_mut() {
+            ss.close().await?;
+        }
+        Ok(())
+    }
 
     /// Handle initial messages.
     /// This has different error handling strategy (loud) vs handle(...), which suppresses errors.
@@ -557,11 +633,17 @@ impl StateHandler {
         }
 
         // Query for the target gene
-        let track_service = state.track_service_checked()?;
+        if state.data.track_service.is_none() {
+            return Err(TGVError::StateError(
+                "Track service not initialized".to_string(),
+            ));
+        }
+
+        let track_service = state.data.track_service.as_mut().unwrap();
         let gene = track_service
             .query_k_genes_after(
                 state.reference_checked()?,
-                state.contig()?,
+                &state.contig()?,
                 middle,
                 n,
                 &mut state.data.track_cache,
@@ -733,7 +815,7 @@ impl StateHandler {
         let exon = track_service
             .query_k_exons_after(
                 state.reference_checked()?,
-                state.contig()?,
+                &state.contig()?.clone(),
                 middle,
                 n,
                 &mut state.data.track_cache,
@@ -854,4 +936,181 @@ impl StateHandler {
             "Failed to find a default initial region. Please provide a starting region with -r.".to_string(),
         ))
     }
+}
+
+
+impl StateHandler {
+
+
+    pub async fn handle_data_messages(
+        &mut self,
+        reference: Option<&Reference>, // TODO: improve this.
+        data_messages: Vec<DataMessage>,
+    ) -> Result<bool, TGVError> {
+        let mut loaded_data = false;
+        for data_message in data_messages {
+            loaded_data = self.handle_data_message(reference, data_message).await?;
+        }
+        Ok(loaded_data)
+    }
+
+    // TODO: async
+    pub async fn handle_data_message(
+        &mut self,
+        reference: Option<&Reference>, // TODO: improve this.
+        data_message: DataMessage,
+    ) -> Result<bool, TGVError> {
+        let mut loaded_data = false;
+
+        match data_message {
+            DataMessage::RequiresCompleteAlignments(region) => {
+
+                if !self.has_complete_alignment(&region) {
+                    self.alignment =Some(self.alignment_repository.read_alignment(&region)?); // TODO
+                    loaded_data = true;
+                }
+            }
+            DataMessage::RequiresCompleteFeatures(region) => {
+                let has_complete_track = self.has_complete_track(&region);
+                if let (Some(reference), Some(track_service)) =
+                    (reference, self.track_service.as_mut())
+                {
+                    if !has_complete_track {
+                        if let Ok(track) = track_service
+                            .query_gene_track(reference, &region, &mut self.track_cache)
+                            .await
+                        {
+                            self.track = Some(track);
+                            loaded_data = true;
+                        } else {
+                            // Do nothing (track not found). TODO: fix this shit properly.
+                        }
+                    }
+                } else if reference.is_none() {
+                    // No reference provided, cannot load features
+                } else {
+                    return Err(TGVError::StateError(
+                        "Track service not initialized".to_string(),
+                    ));
+                }
+            }
+            DataMessage::RequiresCompleteSequences(region) => {
+                if self.sequence_service.is_none() {
+                    return Err(TGVError::IOError("Sequence service not found".to_string()));
+                }
+                let sequence_service = self.sequence_service.as_ref().unwrap();
+
+                if !self.has_complete_sequence(&region) {
+                    match sequence_service.query_sequence(&region).await {
+                        Ok(sequence) => {
+                            self.sequence = Some(sequence);
+                            loaded_data = true;
+                        }
+                        Err(_) => {
+                            return Err(TGVError::IOError("Sequence service error".to_string()));
+                        }
+                    }
+                }
+            }
+
+            DataMessage::RequiresCytobands(contig) => {
+                if self.contigs.cytoband_is_loaded(&contig)? {
+                    return Ok(false);
+                }
+
+                if let (Some(reference), Some(track_service)) =
+                    (reference, self.track_service.as_ref())
+                {
+                    let cytoband = track_service.get_cytoband(reference, &contig).await?;
+                    self.contigs.update_cytoband(&contig, cytoband);
+                    loaded_data = true;
+                } else if reference.is_none() {
+                    // Cannot load cytobands without reference
+                } else {
+                    // track service not available
+                }
+            }
+        }
+
+        Ok(loaded_data)
+    }
+
+    pub async fn load_all_data(
+        &mut self,
+        reference: Option<&Reference>, // TODO: improve this.
+        region: Region,
+    ) -> Result<bool, TGVError> {
+        let loaded_alignment = self
+            .handle_data_message(
+                reference,
+                DataMessage::RequiresCompleteAlignments(region.clone()),
+            )
+            .await?;
+        let loaded_track = self
+            .handle_data_message(
+                reference,
+                DataMessage::RequiresCompleteFeatures(region.clone()),
+            )
+            .await?;
+        let loaded_sequence = self
+            .handle_data_message(
+                reference,
+                DataMessage::RequiresCompleteSequences(region.clone()),
+            )
+            .await?;
+        Ok(loaded_alignment || loaded_track || loaded_sequence)
+    }
+
+    pub fn has_complete_alignment(&self, region: &Region) -> bool {
+        self.alignment.is_some() && self.alignment.as_ref().unwrap().has_complete_data(region)
+    }
+
+    pub fn has_complete_track(&self, region: &Region) -> bool {
+        // self.track_cache.get_track(region.contig()) == Some(None)
+        self.track.is_some() && self.track.as_ref().unwrap().has_complete_data(region)
+    }
+
+    pub fn has_complete_sequence(&self, region: &Region) -> bool {
+        self.sequence.is_some() && self.sequence.as_ref().unwrap().has_complete_data(region)
+    }
+
+    pub fn track_checked(&self) -> Result<&Track<Gene>, TGVError> {
+        self.track.as_ref().ok_or(TGVError::StateError("Track is not initialized".to_string()))
+    }
+}
+
+
+pub async fn load_contig_data(
+    reference: Option<&Reference>,
+    track_service: Option<&TrackServiceEnum>,
+    repository: &AlignmentRepositoryEnum,
+) -> Result<ContigCollection, TGVError> {
+    let mut contig_data = ContigCollection::new(reference.cloned());
+
+    if let (Some(reference), Some(track_service)) = (reference, track_service) {
+        for (contig, length) in track_service.get_all_contigs(reference).await? {
+            contig_data
+                .update_or_add_contig(contig, Some(length))
+                .unwrap();
+        }
+    }
+
+    if let Some(reference) = reference {
+        match &reference {
+            Reference::Hg19 | Reference::Hg38 => {
+                for cytoband in Cytoband::from_human_reference(reference)?.iter() {
+                    contig_data.update_cytoband(&cytoband.contig, Some(cytoband.clone()));
+                }
+            }
+            _ => {}
+        }
+    }
+
+    if  !matches!(repository, AlignmentRepositoryEnum::None) {
+        contig_data
+            .update_from_bam(reference, repository)
+            .unwrap();
+    }
+
+    Ok(contig_data)
 }
